backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.25
pymysql>=1.1.0
python-dotenv>=1.0.1
feedparser>=6.0.11
requests>=2.31.0
playwright>=1.41.2
pydantic>=2.6.1
pydantic-settings>=2.1.0
pandas>=2.2.0
langchain>=0.1.5
chromadb>=0.4.22
schedule>=1.2.1
beautifulsoup4>=4.12.3
langdetect>=1.0.9
langchain-google-genai>=1.0.1
